requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy
flask-cors
openai~=0.27.8
sentence_transformers
torch~=2.2.2
pandas~=1.5.3
selenium~=4.10.0
python-dotenv~=1.0.1
Unidecode~=1.3.8